+++ src/ralph/schemas/edx/converters/xapi/constants.py
@@ -1,0 +1,30 @@
+"""Constants for xAPI specifications"""
+
+# Languages
+EN = "en"
+
+# xAPI activities
+XAPI_ACTIVITY_PAGE = "http://activitystrea.ms/schema/1.0/page"
+
+# xAPI verbs
+XAPI_VERB_VIEWED = "http://id.tincanapi.com/verb/viewed"
+
+# xAPI extensions
+XAPI_EXTENSION_ACCEPT_LANGUAGE = "https://www.edx.org/extension/accept_language"
+XAPI_EXTENSION_AGENT = "https://www.edx.org/extension/agent"
+XAPI_EXTENSION_COURSE_ID = "https://www.edx.org/extension/course_id"
+XAPI_EXTENSION_COURSE_USER_TAGS = "https://www.edx.org/extension/course_user_tags"
+XAPI_EXTENSION_HOST = "https://www.edx.org/extension/host"
+XAPI_EXTENSION_IP = "https://www.edx.org/extension/ip"
+XAPI_EXTENSION_ORG_ID = "https://www.edx.org/extension/org_id"
+XAPI_EXTENSION_PATH = "https://www.edx.org/extension/path"
+XAPI_EXTENSION_REFERER = "https://www.edx.org/extension/referer"
+XAPI_EXTENSION_REQUEST = "https://www.edx.org/extension/request"
+XAPI_EXTENSION_USER_ID = "https://www.edx.org/extension/user_id"
+
+# Display Names
+VIEWED = "viewed"
+PAGE = "page"
+
+# Other
+VERSION = "1.0.3"